requirements.txt
fastapi
uvicorn[standard]
openpyxl
sqlalchemy
pandas
psycopg2-binary
pydantic
pyyaml